Add encrypt command
Adding support for encrypting EJSON files. This command mimics the
baheviour of EJSON by taking one or more files as arguments and
encrypting them in turn.

Within each file, only eligible keys, that is, string values who's keys
are not prefixed with underscore will be encrypted. Values that are
already encrypted will not be altered.