Add support for compacting values
Sometimes it's nice to be able to paste a multiline string into a field
before encrypting it. Multiline values are tricky but sometimes useful
for things like service accounts for example.

In order to support this, eligible keys are _compacted_ before
encryption which simply means they are trimmed of all leading/trailing
whitespace and have any newlines (`\r`, `\n`) escaped.