Setup goreleaser
Adding a bunch of fuckery to make goreleaser handle cross compiled rust
binaries. goreleaser will take care of publishing the docker images and
creating a draft release on GitHub with all of the binaries attached to
it.

Eventually, I'll have the crate published here too and move this into a
GitHub workflow but for now, see README.md for usage details.