Add musl target to releases
When trying to use the linux_amd64 build in alpine you get an error
about missing libraries. Adding a build that targets
`x86_64-unknown-linux-musl` resolves this issue.

Due to the fact that I'm hacking goreleaser to work with cross compiling
rust apps, I need to use an existing architecture that is a combination
of GOOS and GOARCH64. So here, the linux_amd64_v2 architecture is used
to mean "musl build".